fix: u8g2 lookup_glyph 返回 int16_t 溢位导致大字体字符丢失
wqy13 字体数据 225KB，CJK 字符索引偏移超过 32767，
eui_font_u8g2_lookup_glyph 返回 int16_t 造成符号扩展为负数，
被 draw_u8g2_glyph 视为未找到而跳过渲染。

修复: int16_t → int32_t (函数返回, 字形偏移, bitmap 字节位)
同步扩大 bit_reader_t.byte_pos, u8g2_glyph_t.bitmap_byte 为 uint32_t

+++ include/eui/eui_types.h
@@ -60,7 +60,7 @@ typedef struct {
 
 struct eui_font;
 
-typedef int16_t (*eui_font_lookup_fn)(const struct eui_font *font,
+typedef int32_t (*eui_font_lookup_fn)(const struct eui_font *font,
                                        uint16_t encoding, uint16_t prev);
 
 typedef struct eui_font {
